refactor: migrate leaderboard functions to a new module for improved LCP
